+upd: send email
1. Separated components in CartModal
2. Preparation to send email - I got error - I created issues on react-email and next js
