+++ app/api/create-checkout-session/route.ts
@@ -12,21 +12,21 @@ export async function POST(request: Request) {
   const body = (await request.json()) as { stripeProductsQuery: string }
 
   try {
-    //Decoding url to get json from this %7B%22price%22%3A%22price_1O8TYeDEq5VtEmnoi7r1D4Gs%22%2C%22quantity%22%3A1%7D
+    // Decoding url to get json from this %7B%22price%22%3A%22price_1O8TYeDEq5VtEmnoi7r1D4Gs%22%2C%22quantity%22%3A1%7D
     const productsQuery = decodeURIComponent(body.stripeProductsQuery)
 
-    //split & if query has multiple objects to get array from this
-    //{"price":"price_1O8TYeDEq5VtEmnoi7r1D4Gs","quantity":1}&{"price":"price_1O7174DEq5VtEmno3dk9tomc","quantity":1}
+    // split & if query has multiple objects to get array from this
+    // {"price":"price_1O8TYeDEq5VtEmnoi7r1D4Gs","quantity":1}&{"price":"price_1O7174DEq5VtEmno3dk9tomc","quantity":1}
     const productsStringArray = productsQuery.includes("&") ? productsQuery.split("&") : [productsQuery]
 
-    //parsing string to convert string type to json type
+    // Parsing string to convert string type to json type
     const productsJsonArray: stripeProductType[] = productsStringArray.map(product => JSON.parse(product))
 
     const session = await stripe.checkout.sessions.create({
       billing_address_collection: "required",
       line_items: productsJsonArray,
       mode: "payment",
-      success_url: `${getURL()}payment/?status=success`,
+      success_url: `${getURL()}payment/?status=success&session_id={CHECKOUT_SESSION_ID}`,
       cancel_url: `${getURL()}payment/?status=canceled`,
       shipping_address_collection: {
         allowed_countries: ["DE"],
@@ -34,7 +34,7 @@ export async function POST(request: Request) {
     })
 
     if (session.url) {
-      return NextResponse.json(session.url)
+      return NextResponse.json(decodeURIComponent(session.url))
     } else {
       return new NextResponse("No session url - please check create-checkout-session route", { status: 500 })
     }
