--- app/api/email/route.ts
@@ -1,5 +1,0 @@
-import { NextResponse } from "next/server"
-
-export async function POST() {
-  return NextResponse.json({ willbedeleted: "123" })
-}
