+++ app/libs/resend.ts
@@ -1,0 +1,3 @@
+import { Resend } from "resend"
+
+export const resend = new Resend(process.env.REACT_RESEND_PUBLIC)
