+++ package.json
@@ -15,35 +15,36 @@
     "@react-email/button": "^0.0.10",
     "@react-email/components": "^0.0.11",
     "@react-email/html": "^0.0.4",
+    "@react-email/render": "0.0.10",
     "@react-email/tailwind": "0.0.8",
+    "resend": "^2.1.0",
+    "react-email": "^1.10.0",
     "@sindresorhus/slugify": "^2.2.1",
-    "@stripe/stripe-js": "^2.2.0",
+    "@stripe/stripe-js": "^2.3.0",
     "@supabase/auth-helpers-nextjs": "^0.8.7",
-    "@supabase/supabase-js": "^2.38.5",
+    "@supabase/supabase-js": "^2.39.3",
     "autoprefixer": "10.4.16",
-    "axios": "^1.6.2",
+    "axios": "^1.6.5",
     "class-variance-authority": "^0.7.0",
     "disposable-email-domains": "^1.0.62",
-    "framer-motion": "^10.16.5",
+    "framer-motion": "^10.18.0",
     "lodash": "^4.17.21",
-    "next": "^14.0.3",
+    "next": "^14.0.4",
     "postcss": "8.4.31",
     "pusher": "^5.2.0",
     "pusher-js": "^8.3.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-email": "^1.9.5",
-    "react-hook-form": "^7.48.2",
+    "react-hook-form": "^7.49.3",
     "react-icons": "^4.12.0",
     "react-images-uploading": "^3.1.7",
     "react-loading-skeleton": "^3.3.1",
     "react-responsive-carousel": "^3.2.23",
     "react-swipeable": "^7.0.1",
-    "resend": "^2.0.0",
-    "stripe": "^14.5.0",
+    "stripe": "^14.12.0",
     "tailwind-merge": "^1.14.0",
     "uuid": "^9.0.1",
-    "zustand": "^4.4.6"
+    "zustand": "^4.4.7"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.202",
@@ -53,8 +54,8 @@
     "@types/uuid": "^9.0.7",
     "eslint": "8.49.0",
     "eslint-config-next": "13.5.2",
-    "pnpm": "^8.10.5",
-    "prettier": "^3.1.0",
+    "pnpm": "^8.14.1",
+    "prettier": "^3.2.2",
     "tailwindcss": "3.3.3",
     "typescript": "5.2.2"
   }
